Computation model & types and functions

+++ src/main/kotlin/de/krall/spreadsheets/value/EvaluatedValue.kt
@@ -1,10 +1,8 @@
 package de.krall.spreadsheets.value
 
-import java.math.BigDecimal
-
 sealed class EvaluatedValue {
     data class Text(val text: String) : EvaluatedValue()
-    data class Number(val number: BigDecimal) : EvaluatedValue()
+    data class Number(val number: Double) : EvaluatedValue()
     data object Unevaluated : EvaluatedValue()
     data object BadFormula : EvaluatedValue()
     data object CircularDependencies : EvaluatedValue()
