+++ src/main/kotlin/de/krall/spreadsheets/value/parser/tree/SlNumberValue.kt
@@ -4,7 +4,7 @@ import de.krall.spreadsheets.value.parser.SlSource
 import java.math.BigDecimal
 
 class SlNumberValue(
-    val number: BigDecimal,
+    val number: Double,
     override val source: SlSource? = null,
 ) : SlValue() {
